simplify example script
